+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Note: Numbers like (#123) point to closed Pull Requests on the fractal-vizarr-vi
 
 # Unreleased
 
+* Used BioNGFF Viewer instead of Vizarr (\#69).
 * Added workaround to fix labels bug in most recent Chrome versions (\#68);
 
 # 0.4.0